+++ src/django_components/__init__.py
@@ -4,7 +4,7 @@
 # NOTE: Middleware is exposed via django_components.middleware
 # NOTE: Some of the documentation is generated based on these exports
 # isort: off
-from django_components.app_settings import ContextBehavior
+from django_components.app_settings import ContextBehavior, ComponentsSettings
 from django_components.autodiscovery import autodiscover, import_libraries
 from django_components.component import Component, ComponentVars, ComponentView
 from django_components.component_registry import (
@@ -39,6 +39,7 @@ __all__ = [
     "autodiscover",
     "cached_template",
     "ContextBehavior",
+    "ComponentsSettings",
     "Component",
     "ComponentFormatter",
     "ComponentRegistry",